Content Model Shadow Edit Step 2: Split editor
