+++ packages/roosterjs-editor-core/lib/index.ts
@@ -1,2 +1,3 @@
 // Classes
 export { default as Editor } from './editor/Editor';
+export { EditorBase } from './editor/EditorBase';
